pyproject.toml
[project]
name = "data-decide"
version = "0.1.0"
description = "OLMo training and evaluation framework using DataDecide methodology"
authors = [
    { name="FinPile Team", email="finpile@gatech.edu" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
keywords = ["olmo", "language-model", "datadecide", "ml", "evaluation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core ML frameworks
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "accelerate>=0.20.0",

    # Training utilities
    "wandb>=0.15.0",
    "tensorboard>=2.13.0",
    "rich>=13.0.0",

    # Data processing
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",

    # Configuration
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",

    # System monitoring
    "psutil>=5.9.0",

    # Tokenization
    "tokenizers>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]
telemetry = [
    "gputil>=1.4.0",
    "psutil>=5.9.0",
]

[project.scripts]
data-decide-train = "data_decide.scripts.train:main"
data-decide-analyze = "data_decide.scripts.analyze_run:main"
data-decide-monitor = "data_decide.scripts.monitor_training:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["data_decide*"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long - handled by formatter

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__ files are ok
